+++ libdariadb/storage/memstorage/memstorage.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <libdariadb/interfaces/ichunkcontainer.h>
+#include <libdariadb/storage/chunkcontainer.h>
 #include <libdariadb/interfaces/imeasstorage.h>
 #include <libdariadb/meas.h>
 #include <libdariadb/st_exports.h>
@@ -30,7 +30,9 @@ public:
   EXPORT virtual Time maxTime() override;
   EXPORT virtual bool minMaxTime(dariadb::Id id, dariadb::Time *minResult,
                                  dariadb::Time *maxResult) override;
-  EXPORT virtual void foreach (const QueryInterval &q, IReaderClb * clbk) override;
+  EXPORT virtual Id2Cursor intervalReader(const QueryInterval &q)override;
+  EXPORT  Statistic stat(const Id id, Time from, Time to)override;
+  EXPORT virtual void foreach (const QueryInterval &q, IReadCallback * clbk) override;
   EXPORT virtual Id2Meas readTimePoint(const QueryTimePoint &q) override;
   EXPORT virtual Id2Meas currentValue(const IdArray &ids, const Flag &flag) override;
   using IMeasStorage::append;