+++ Changelog.md
@@ -1,3 +1,16 @@
+v0.3.0
+=====
+- Remove bystep storage engine.
+- Accept unordered data.
+- Support join of queries.
+- Statistic:
+ - time min/max
+ - value min/max
+ - measurement count
+ - values sum
+- Page can be repacked.
+- Scheme API.
+
 v0.2.0
 =====
 Server: